requirements.txt
fastapi==0.116.1
pyrogram==2.0.106
tgcrypto==1.2.5
python-dotenv==1.1.1
uvicorn==0.35.0
